requirements.txt
streamlit
python-dotenv
langchain-community
langchain
requests
pillow
pymysql
sqlalchemy
